Move fpn to fpn module
