+++ src/netspresso_trainer/models/necks/__init__.py
@@ -1,0 +1,1 @@
+from .experimental.fpn import fpn
